+++ src/interfaces.ts
@@ -1,29 +1,5 @@
-import { TFile, TFolder, View } from 'obsidian';
 import { ProviderType } from './constants';
 
 export interface Settings {
-  commands: Record<ProviderType | 'global', boolean>;
-  folderContextMenu: boolean;
-  expandAttachmentFolder: boolean;
-  splitButtons: boolean;
-}
-
-export interface FileExplorerItem {
-  file: TFile | TFolder;
-  collapsed?: boolean;
-  setCollapsed?: (state: boolean) => void;
-}
-
-export interface TagExplorerItem {
-  tag: string;
-  children?: TagExplorerItem[];
-  collapsed?: boolean;
-  setCollapsed?: (state: boolean) => void;
-  vChildren?: {
-    children: TagExplorerItem[]
-  };
-}
-
-export interface TagPaneView extends View {
-  tagDoms: TagExplorerItem[]
+  commands: Record<ProviderType, boolean>;
 }