+++ src/provider/index.ts
@@ -1,2 +1,3 @@
 export { FileExplorerProvider } from './file-explorer';
 export { TagPaneProvider } from './tag-pane';
+export { GlobalProvider } from './global';